requirements.txt
certifi==2017.4.17
chardet==3.0.4
Django==1.11.2
django-filter==1.0.4
djangorestframework==3.6.3
idna==2.5
Markdown==2.6.8
mysqlclient==1.3.10
pytz==2017.2
requests==2.18.1
urllib3==1.21.1
django-rest-swagger==2.1.2